requirements.txt
fastapi[all]
sqlalchemy
pydantic
uvicorn
strawberry-graphql[cli]